requirements.txt
google-cloud-aiplatform
fastapi
uvicorn[standard]
pytest
pydantic
python-dotenv